+++ src/app/api/api-injection-tokens.ts
@@ -7,13 +7,11 @@
 
 import { InjectionToken } from "@angular/core";
 import { BillingType } from "src/models/entities/BillingType";
-import { Customer } from "src/models/entities/Customer";
 import { Image } from "src/models/entities/Image";
 import { Person } from "src/models/entities/Person";
 import { Product } from "src/models/entities/Product";
 import { ProductCategory } from "src/models/entities/ProductCategory";
 import { ProductList } from "src/models/entities/ProductList";
-import { Salesperson } from "src/models/entities/Salesperson";
 import { Sell } from "src/models/entities/Sell";
 import { Shipper } from "src/models/entities/Shipper";
 import { User } from "src/models/entities/User";
@@ -36,14 +34,14 @@ export const API_INJECTION_TOKENS = {
   access: new InjectionToken<IAccessApiService>('Trebol.IAccessApi'),
   accountProfile: new InjectionToken<IProfileAccountApiService>('Trebol.IProfileAccountApi'),
   dataBillingTypes: new InjectionToken<IEntityDataApiService<BillingType>>('Trebol.IEntityDataApi<BillingType>'),
-  dataCustomers: new InjectionToken<ITransactionalEntityDataApiService<Customer>>('Trebol.ITransactionalEntityDataApiService<Customer>'),
+  dataCustomers: new InjectionToken<ITransactionalEntityDataApiService<Person>>('Trebol.ITransactionalEntityDataApiService<Customer>'),
   dataImages: new InjectionToken<ITransactionalEntityDataApiService<Image>>('Trebol.ITransactionalEntityDataApiService<Image>'),
   dataPeople: new InjectionToken<IEntityDataApiService<Person>>('Trebol.IEntityDataApi<Person>'),
   dataProductCategories: new InjectionToken<ITransactionalEntityDataApiService<ProductCategory>>('Trebol.ITransactionalEntityDataApiService<ProductCategory>'),
   dataProductLists: new InjectionToken<ITransactionalEntityDataApiService<ProductList>>('Trebol.ITransactionalEntityDataApiService<ProductList>'),
   dataProducts: new InjectionToken<ITransactionalEntityDataApiService<Product>>('Trebol.ITransactionalEntityDataApiService<Product>'),
   dataSales: new InjectionToken<ITransactionalEntityDataApiService<Sell>>('Trebol.ITransactionalEntityDataApiService<Sell>'),
-  dataSalespeople: new InjectionToken<ITransactionalEntityDataApiService<Salesperson>>('Trebol.ITransactionalEntityDataApiService<Salesperson>'),
+  dataSalespeople: new InjectionToken<ITransactionalEntityDataApiService<Person>>('Trebol.ITransactionalEntityDataApiService<Salesperson>'),
   dataShippers: new InjectionToken<ITransactionalEntityDataApiService<Shipper>>('Trebol.ITransactionalEntityDataApiService<Shipper>'),
   dataUserRoles: new InjectionToken<IEntityDataApiService<UserRole>>('Trebol.IEntityDataApi<UserRole>'),
   dataUsers: new InjectionToken<ITransactionalEntityDataApiService<User>>('Trebol.ITransactionalEntityDataApiService<User>'),