+++ src/app/api/local-memory/data/salespeople-data.local-memory-api.service.ts
@@ -7,14 +7,14 @@
 
 import { Injectable } from '@angular/core';
 import { compareObjectsForSort } from 'src/functions/compareObjectsForSort';
-import { Salesperson } from 'src/models/entities/Salesperson';
+import { Person } from 'src/models/entities/Person';
 import { matchesStringProperty, matchesNumberProperty } from '../entity-data.local-memory-api.functions';
 import { MOCK_SALESPEOPLE } from '../mock/mock-salespeople.datasource';
 import { TransactionalEntityDataLocalMemoryApiService } from '../transactional-entity-data.local-memory-api.abstract.service';
 
 @Injectable()
 export class SalespeopleDataLocalMemoryApiService
-  extends TransactionalEntityDataLocalMemoryApiService<Salesperson> {
+  extends TransactionalEntityDataLocalMemoryApiService<Person> {
 
   protected items = MOCK_SALESPEOPLE.slice();
 
@@ -26,17 +26,17 @@ export class SalespeopleDataLocalMemoryApiService
    * Iterates each key-value property pair in the provided object,
    * filters all items matching the same properties.
    */
-  protected filterItems(filter: any): Salesperson[] {
+  protected filterItems(filter: any): Person[] {
     let matchingItems = this.items;
     for (const propName in filter) {
       if (filter.hasOwnProperty(propName) && propName !== 'id') {
         const propValue = filter[propName];
         if (typeof propValue === 'string') {
-          matchingItems = matchingItems.filter(it => matchesStringProperty(it.person, propName, propValue));
+          matchingItems = matchingItems.filter(it => matchesStringProperty(it, propName, propValue));
         } else if (typeof propValue === 'number') {
-          matchingItems = matchingItems.filter(it => matchesNumberProperty(it.person, propName, propValue));
+          matchingItems = matchingItems.filter(it => matchesNumberProperty(it, propName, propValue));
         } else if (propValue === null) {
-          matchingItems = matchingItems.filter(it => (it.person[propName] === null));
+          matchingItems = matchingItems.filter(it => (it[propName] === null));
         }
       }
     }
@@ -44,18 +44,18 @@ export class SalespeopleDataLocalMemoryApiService
     return matchingItems;
   }
 
-  protected sortItems(a: Salesperson, b: Salesperson, sortBy: string, order = 'asc'): number {
+  protected sortItems(a: Person, b: Person, sortBy: string, order = 'asc'): number {
     const objectSortProperty =
       (sortBy === 'name') ? 'lastName' :
       sortBy;
-    return compareObjectsForSort(a.person, b.person, objectSortProperty, order);
+    return compareObjectsForSort(a, b, objectSortProperty, order);
   }
 
-  protected itemExists(salesperson: Partial<Salesperson>) {
-    return this.items.some(salesperson2 => (salesperson.person.idNumber === salesperson2.person.idNumber));
+  protected itemExists(salesperson: Partial<Person>) {
+    return this.items.some(salesperson2 => (salesperson.idNumber === salesperson2.idNumber));
   }
 
-  protected getIndexOfItem(customer: Partial<Salesperson>) {
-    return this.items.findIndex(salesperson2 => (customer.person.idNumber === salesperson2.person.idNumber));
+  protected getIndexOfItem(customer: Partial<Person>) {
+    return this.items.findIndex(salesperson2 => (customer.idNumber === salesperson2.idNumber));
   }
 }