+++ src/app/api/local-memory/mock/mock-salespeople.datasource.ts
@@ -5,23 +5,13 @@
  * https://opensource.org/licenses/MIT
  */
 
-import { Salesperson } from 'src/models/entities/Salesperson';
 import { MOCK_PEOPLE } from './mock-people.datasource';
+import { Person } from 'src/models/entities/Person';
 
-export const MOCK_SALESPEOPLE: Salesperson[] = [
-  {
-    person: MOCK_PEOPLE[4]
-  },
-  {
-    person: MOCK_PEOPLE[5]
-  },
-  {
-    person: MOCK_PEOPLE[6]
-  },
-  {
-    person: MOCK_PEOPLE[7]
-  },
-  {
-    person: MOCK_PEOPLE[8]
-  }
+export const MOCK_SALESPEOPLE: Person[] = [
+  MOCK_PEOPLE[4],
+  MOCK_PEOPLE[5],
+  MOCK_PEOPLE[6],
+  MOCK_PEOPLE[7],
+  MOCK_PEOPLE[8]
 ];