+++ src/app/management/routes/customers/management-customers.service.ts
@@ -8,15 +8,15 @@
 import { Inject, Injectable } from '@angular/core';
 import { API_INJECTION_TOKENS } from 'src/app/api/api-injection-tokens';
 import { IEntityDataApiService } from 'src/app/api/entity.data-api.iservice';
-import { Customer } from 'src/models/entities/Customer';
 import { DataManagerServiceDirective } from '../../directives/data-manager/data-manager.service.directive';
+import { Person } from 'src/models/entities/Person';
 
 @Injectable()
 export class ManagementCustomersService
-  extends DataManagerServiceDirective<Customer> {
+  extends DataManagerServiceDirective<Person> {
 
   constructor(
-    @Inject(API_INJECTION_TOKENS.dataCustomers) protected dataService: IEntityDataApiService<Customer>
+    @Inject(API_INJECTION_TOKENS.dataCustomers) protected dataService: IEntityDataApiService<Person>
   ) {
     super();
   }