+++ src/app/management/routes/salespeople/management-salespeople.component.spec.ts
@@ -18,7 +18,7 @@ import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 import { RouterTestingModule } from '@angular/router/testing';
 import { EMPTY, Observable, of } from 'rxjs';
 import { MOCK_SALESPEOPLE } from 'src/app/api/local-memory/mock/mock-salespeople.datasource';
-import { Salesperson } from 'src/models/entities/Salesperson';
+import { Person } from 'src/models/entities/Person';
 import { EntityFormDialogComponent } from '../../dialogs/entity-form/entity-form-dialog.component';
 import { EntityFormDialogConfig } from '../../dialogs/entity-form/EntityFormDialogConfig';
 import { ManagementSalespeopleComponent } from './management-salespeople.component';
@@ -119,7 +119,7 @@ describe('ManagementSalespeopleComponent', () => {
           apiService: serviceSpy.dataService
         },
         width: '40rem'
-      } as EntityFormDialogConfig<Salesperson>
+      } as EntityFormDialogConfig<Person>
     );
   });
 });