+++ CHANGELOG.md
@@ -7,6 +7,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Adapting to Trebol API v3
+  - Remove classes `Customer` and `Salesperson` in favor of `Person`
+    - As a result, in both cases, reference tokens to personal information are shorter
+      - `customer.person.firstName` -> `customer.firstName`
+
 ## [v2.12.1-rc.3] - 2022-08-02
 
 ### Changed